chore: extract createNodes from fx-fore and add unit tests
Also remove a second dead implementation of createNodes. One is enough.
