+++ test/xpathutil.test.js
@@ -55,29 +55,4 @@ describe('XPathUtil Tests', () => {
     const action = el.querySelector('fx-setvalue');
     expect(XPathUtil.getParentBindingElement(action)).to.equal(inner);
   });
-
-  describe('createNodesFromXPath', () => {
-    const document = new window.DOMParser().parseFromString('<xml/>', 'text/xml');
-    it(`works for "listBibl[@type='transmission']/bibl"`, () => {
-      const el = fixtureSync(html`
-        <fx-fore>
-          <fx-model>
-            <fx-instance>
-              <data>
-                <counter>0</counter>
-              </data>
-            </fx-instance>
-          </fx-model>
-        </fx-fore>
-      `);
-
-      const node = XPathUtil.createNodesFromXPath(
-        `listBibl[@type='transmission']/bibl`,
-        document,
-        el,
-      );
-
-      expect(node.outerHTML).to.equal('<listBibl type="transmission"><bibl/></listBibl>');
-    });
-  });
 });
